LEFT MOST CHARACTER THAT DOES NOT REPEAT
Important for coding Interviews. If index[s1[i]]==-1, we assign the value of index. Else we assign some other value. So if it repeats we can avoid that. Then we traverse all the elements in the findIndex array and compare with res and return the minimum. If all characters are repeating, res i.e -1 is returned